+++ docs/fr_FR/Changelog.md
@@ -3,6 +3,7 @@
 - Amélioration: Suppression cmde obsolete 'Up-Down' => 'inf_zbCmdC-0008-MoveToLevel'.
 - Correction: Interne CmdD (configureZigate()).
 - Amélioration: Ajout version Jeedom dans les infos clefs.
+- Amélioration: Interne AbeilleCmdQueue: processCmdQueues() revu.
 
 ## 260705-BETA-2, 260705-BETA-1